Split Scramble Component
